Content Model: Reorganize some code
