+++ packages-content-model/roosterjs-content-model-dom/lib/modelApi/common/hasSpacesOnlyTest.ts
@@ -1,4 +1,4 @@
-import { hasSpacesOnly } from '../../lib/domUtils/stringUtil';
+import { hasSpacesOnly } from '../../lib/modelApi/common/hasSpacesOnly';
 
 describe('hasSpacesOnly', () => {
     it('Empty string', () => {
