+++ packages-content-model/roosterjs-content-model-editor/lib/domUtils/stringUtil.ts
@@ -1,11 +1,8 @@
-// A regex to match text that only has space and CR
-// We use real space char " " (\u0020) here but not "\s" since "\s" will also match "&nbsp;" (\u00A0) which is something we need to keep
-const SPACE_TEXT_REGEX = /^[\r\n\t ]*$/;
-
 const SPACES_REGEX = /[\u2000\u2009\u200a​\u200b​\u202f\u205f​\u3000\s\t\r\n]/gm;
 const PUNCTUATIONS = '.,?!:"()[]\\/';
 
 /**
+ * @internal
  * Check if the given character is punctuation
  * @param char The character to check
  */
@@ -14,6 +11,7 @@ export function isPunctuation(char: string) {
 }
 
 /**
+ * @internal
  * Check if the give character is a space. A space can be normal ASCII pace (32) or non-break space (160) or other kinds of spaces
  * such as ZeroWidthSpace, ...
  * @param char The character to check
@@ -25,14 +23,6 @@ export function isSpace(char: string) {
 
 /**
  * @internal
- * Check if the given string only has space, including line breaks.
- * @param txt The string to check
- */
-export function hasSpacesOnly(txt: string): boolean {
-    return SPACE_TEXT_REGEX.test(txt);
-}
-
-/**
  * Normalize spaces of the given string. After normalization, all leading (for forward) or trailing (for backward) spaces
  * will be replaces with non-break space (160)
  * @param txt The string to normalize
